src/test/java/com/karakays/leetcode/solutions/S240Test.java
package com.karakays.leetcode.solutions;

import com.karakays.leetcode.solutions.S19.ListNode;
import com.karakays.leetcode.utils.LinkedListUtils;
import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

public class S240Test {

    S240 s240 = new S240();

    @Test
    public void test1() {
        int[][] matrix = {
  {1,   4,  7, 11, 15},
  {2,   5,  8, 12, 19},
  {3,   6,  9, 16, 22},
  {10, 13, 14, 17, 24},
  {18, 21, 23, 26, 30}
};
        Assert.assertEquals(true, s240._searchMatrix(matrix, 5));
    }

    @Test
    public void test2() {
        int[][] matrix = {
                {1,   4,  7, 11, 15},
                {2,   5,  8, 12, 19},
                {3,   6,  9, 16, 22},
                {10, 13, 14, 17, 24},
                {18, 21, 23, 26, 30}
        };
        Assert.assertEquals(false, s240._searchMatrix(matrix, 20));
    }

}
